+++ app/shared/src/androidMain/kotlin/ui/subject/details/SubjectDetailsPage.android.kt
@@ -3,7 +3,6 @@ package me.him188.ani.app.ui.subject.details
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.Preview
-import me.him188.ani.app.ui.foundation.PreviewData
 import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
 
 @Preview
@@ -11,7 +10,7 @@ import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
 internal fun PreviewSubjectDetails() {
     ProvideCompositionLocalsForPreview {
         val vm = remember {
-            SubjectDetailsViewModel(PreviewData.SOSOU_NO_FURILEN_SUBJECT_ID)
+            SubjectDetailsViewModel(400602)
         }
         SubjectDetailsPage(vm)
     }
