+++ app/shared/src/commonMain/kotlin/ui/foundation/widgets/OneshotTip.kt
@@ -1,4 +1,4 @@
-package me.him188.ani.app
+package me.him188.ani.app.ui.foundation.widgets
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.wrapContentHeight
